Fixed sort_ro tests
